requirements.txt
requests-oauthlib
google-generativeai
python-dotenv